rename participates_helpers.js to cross_frame_helpers.js
