+++ app/javascript/editorV2/panels/condition_preview/forward_proposition/cross_frame/mechanisms/cross_frame_helpers.js
