+++ app/javascript/editorV2/panels/condition_preview/forward_proposition/cross_frame/mechanisms/moved_piece_captures_relation_participant.js
@@ -8,7 +8,7 @@ import { activeAttackOrDefendSets } from '../../relations/attack_or_defend'
 import { activeAdjacentSets } from '../../relations/adjacent'
 import { activeShieldSets } from '../../relations/shield'
 import { singularPosition } from '../../relations/relation_helpers'
-import { singularSquare, commitPriorRegion } from './participates_helpers'
+import { singularSquare, commitPriorRegion } from './cross_frame_helpers'
 import { legalOriginCandidates } from './shifts_mobility_helpers'
 import { committedSpecies } from 'editorV2/panels/condition_preview/shared/singular_constraints'
