+++ app/javascript/editorV2/panels/condition_preview/forward_proposition/cross_frame/mechanisms/moved_piece_obstructs_shield.js
@@ -6,7 +6,7 @@ import {
 } from 'editorV2/panels/condition_preview/shared/geometry_utils'
 import {
   singularSquare, ensureRolePieceAt, commitPriorRegion
-} from './participates_helpers'
+} from './cross_frame_helpers'
 import { roleForPlan } from '../../moved_binding'
 import { committedSpecies } from 'editorV2/panels/condition_preview/shared/singular_constraints'
