+++ app/javascript/editorV2/panels/condition_preview/forward_proposition/cross_frame/mechanisms/moved_piece_participates_in_attack_or_defend.js
@@ -9,7 +9,7 @@ import { placePiece } from 'editorV2/panels/condition_preview/shared/piece_place
 import {
   singularSquare, placeableSpecies, ensureRolePieceAt, commitPriorRegion,
   otherSidePropositionFor
-} from './participates_helpers'
+} from './cross_frame_helpers'
 import { roleForPlan } from '../../moved_binding'
 import { committedSpecies } from 'editorV2/panels/condition_preview/shared/singular_constraints'
