+++ app/javascript/editorV2/panels/condition_preview/forward_proposition/cross_frame/mechanisms/moved_piece_shifts_allied_mobility.js
@@ -6,7 +6,7 @@ import { placeKingDeliberately } from 'editorV2/panels/condition_preview/shared/
 import { placePiece } from 'editorV2/panels/condition_preview/shared/piece_placement'
 import { mobilityAt } from 'gameplay/mobility'
 import { respectsAllCaps } from '../../respect_caps'
-import { singularSquare, commitPriorRegion, entryConcernsMovedPiece } from './participates_helpers'
+import { singularSquare, commitPriorRegion, entryConcernsMovedPiece } from './cross_frame_helpers'
 import {
   legalOriginCandidates, hypotheticalMobilityAt, directionSatisfied
 } from './shifts_mobility_helpers'
