+++ app/javascript/editorV2/panels/condition_preview/forward_proposition/cross_frame/mechanisms/moved_piece_shifts_enemy_king_mobility.js
@@ -8,7 +8,7 @@ import { adjacentNeighborPositions } from 'editorV2/panels/condition_preview/sha
 import { mobilityAt } from 'gameplay/mobility'
 import {
   singularSquare, commitPriorRegion, ensureRolePieceAt
-} from './participates_helpers'
+} from './cross_frame_helpers'
 import {
   legalOriginCandidates, hypotheticalMobilityAt, directionSatisfied,
   enemyKingPosition, ensureEnemyKingPlaced
